requirements.txt
starlette
pdf2docx
docx2pdf
pypdf
pdf2image
img2pdf
python-pptx
pytesseract
Pillow
pillow-heif
cairosvg
ffmpeg-python
mutagen
weasyprint
markdown
openpyxl
pandas
pyzipper
celery
redis
fastapi
uvicorn
python-multipart
aiofiles
pydantic-settings
stripe
PyJWT
httpx
python-jose[cryptography]
sqlmodel
alembic
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
groq


aiosqlite
jinja2